requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-multipart>=0.0.9
boto3>=1.34.0
aiofiles>=23.2.1
pandas>=2.2.0
openpyxl>=3.1.2
reportlab>=4.1.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.5
alembic>=1.13.1
